Add a websocket_cors function
As CORS restrictions don't apply to WebSockets it is important that
the server check the origin is allowed before deciding to accept the
WebSocket. This is made easier via the websocket_cors function which
can be used as so,

    @app.websocket("/ws")
    @websocket_cors(allow_origin="https://quart.com")
    async def ws():
        ...

which will block requests with a 400 response that do not have
`https://quart.com` as the origin header value.